% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivCut.R
\name{equivCut}
\alias{equivCut}
\title{Drop duplicated matched nodes}
\usage{
equivCut(tree1, tree2, id.results, least.inclusive)
}
\arguments{
\item{tree1}{The first tree.}

\item{tree2}{The second tree.}

\item{id.results}{Results from a run of the idMatches function.}

\item{least.inclusive}{Do you want to keep the smallest or largest set of
matched taxa. See details.}
}
\description{
Identify which of the duplicated nodes you want to retain.
}
\details{
The point of this function is to drop duplicated nodes that appear more than once
in the results for tree 2. If there are multiple nodes in tree 2 which could potentially be mapped to
a node in tree 1, do you want to keep the smallest or the largest one? for example:
((A,B),Z) vs (((A,B),Z),C), where C doesn't occur in tree 1. Do you want the matched node
from tree 2 to be A,B,Z or A,B,Z,C.
}
\author{
Eliot Miller
}
