% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoords.R
\name{getCoords}
\alias{getCoords}
\title{Get coordinates for plotting tree}
\usage{
getCoords(tree)
}
\arguments{
\item{tree}{A phylogeny in ape format.}
}
\value{
A list providing details on the plotted tree structure.
}
\description{
Should be an unexported utility function but I'm too lazy.
}
\author{
Eliot Miller and Bruce Martin
}
