% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phy}
\alias{phy}
\title{A phylogeny in ape format.}
\format{Newick format}
\source{
\url{https://github.com/thej022214/corHMM/}
}
\usage{
phy
}
\description{
A phylogeny used for examples in corHMM and shiftPlot.
}
\keyword{datasets}
