% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Precur_res.corHMM}
\alias{Precur_res.corHMM}
\title{corHMM output.}
\format{
Standard corHMM output
}
\source{
\url{https://github.com/thej022214/corHMM/}
}
\usage{
Precur_res.corHMM
}
\description{
The result of following the precursor model example in the corHMM vignette.
}
\keyword{datasets}
