% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processMaps.R
\name{processMaps}
\alias{processMaps}
\title{Process phytools simmaps}
\usage{
processMaps(simmaps, translation.table)
}
\arguments{
\item{simmaps}{A list of simmaps; the result of a call to phytools::make.simmap
where nsim > 1.}

\item{translation.table}{A data frame with the following two columns: "state" and "color",
the name of the color to use in the plot to indicate that that trait state. Color can be
specified in a variety of ways--word, hexadecimal, probably rgb but untested.}
}
\value{
A new states.df.
}
\description{
Convert a list of phytools simmaps to a states.df
}
\details{
Take a list of simmaps, summarize them, and converts the results into a
states.df suitable for downstream use in shiftPlot.
}
\author{
Eliot Miller
}
