% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableSetup.R
\name{tableSetup}
\alias{tableSetup}
\title{Set up a table for node-matching}
\usage{
tableSetup(tree1, tree2)
}
\arguments{
\item{tree1}{The first tree.}

\item{tree2}{The second tree.}
}
\description{
Set up a table for node-matching between two different trees.
}
\details{
Some things.
}
\author{
Eliot Miller
}
