% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processMaps.R
\name{processMaps}
\alias{processMaps}
\title{Process phytools simmaps}
\usage{
processMaps(simmaps, translation.table)
}
\arguments{
\item{simmaps}{A list of simmaps; the result of a call to phytools::make.simmap
where nsim > 1.}

\item{translation.table}{A data frame with the following three columns: orig, the original
names of the states, e.g. "present" or "1,R1"; new, numbers used as shorthand to refer back
to these original state names; color, the name of the color to use in the plot to indicate
that trait state.}
}
\value{
A new states.df.
}
\description{
Convert a list of phytools simmaps to a states.df
}
\details{
Take a list of simmaps, summarize them, and converts the results into a
states.df suitable for downstream use in shiftPlot.
}
\author{
Eliot Miller
}
