% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trianglePlotter.R
\name{trianglePlotter}
\alias{trianglePlotter}
\title{Plot a phylogeny with triangles representing clades}
\usage{
trianglePlotter(
  tree,
  dropped.results,
  clade.table,
  branches,
  presence.color,
  absence.color,
  label.offset,
  text.cex
)
}
\arguments{
\item{tree}{A phylogeny in ape format. This should be the polytomy tree, i.e. the
result of a call to polytomyBind.}

\item{dropped.results}{The result of a call to dropManyTips.}

\item{clade.table}{Data frame with four columns: node, present (0/1), collapse (0/1),
and clade. Node provides values indicating which nodes you will collapse (all nodes
tipwards from the indicated node will be collapsed). Present indicates whether the
tips being collapsed do or do not have the trait in question (I think this column is
not actually used and can be skipped). Collapse indicates whether to actually collapse
that node or not (allowing a user to manually override the results from optimalCollapse).
Clade provides a character string which will be used to rename the collapsed clade.
Initially, this is given a generic name based on the node being collapsed, but these can
be replaced with a name of the user's choosing. This input (clade.table) is the output
of optimalCollapse.}

\item{branches}{Result of a call to firstBranches.}

\item{presence.color}{What color clades for which the trait is present should be colored.}

\item{absence.color}{What color clades for which the trait is absent should be colored.}

\item{label.offset}{How far away from the tips the clade name should be positioned.}

\item{text.cex}{The size of the clade labels.}
}
\description{
Use the optimal collapses to create a tree where triangles represent clades.
}
\details{
Should change the graphical parameters so there's more space for long clade names.
Should not plot the tips. Should allow customization of edge colors.
}
\examples{
#start with a corHMM output and build up a states.df
#load data. these are the results of following the corHMM precursor model vignette
data(Precur_res.corHMM)
data(phy)
nodeStates <- data.frame(present=Precur_res.corHMM$states[,3]+Precur_res.corHMM$states[,4])
tipStates <- data.frame(present=Precur_res.corHMM$tip.states[,3]+Precur_res.corHMM$tip.states[,4])

#note that tip states comes first here!
states <- rbind(tipStates, nodeStates)

#binarize this. choosing to call 0.5 chance of having trait present
states$present[states$present >= 0.5] <- 1
states$present[states$present < 0.5] <- 0

#flip node 103 and all nodes towards tips from there to trait = absent
induced <- states
induced[geiger:::.get.descendants.of.node(103, phy),] <- 0
induced[103,] <- 0

#get rid of row names
row.names(induced) <- NULL

#find the optimal collapse configuration
result <- optimalCollapse(phy, induced, FALSE)

#collapse the tree
dropped <- dropManyTips(phy, result)

#use the polytomyBind function
polyTree <- polytomyBind(dropped, result)

branchingResults <- firstBranches(phy, dropped, result)

saveMe <- trianglePlotter(polyTree, dropped, result, branchingResults, "red", "black", 0.05, 0.2)
}
\author{
Eliot Miller and Bruce Martin
}
